move helper functions to general /utility
